+++ Language/ast/Ast_Struct.cs
@@ -2,49 +2,71 @@
 // Copyright (c) 2021 DiBiAsi Software.
 // Licensed under the Mozila Public License, version 2.0.
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Language
 {
-    public class Ast_Struct: Ast_Base
+    public class Ast_Struct : Ast_Base
     {
         public Ast_Scope StructScope;
         public string Name { get => Token.Lexeme; }
-        public string StructType { get; private set;}
+        public string StructType { get; private set; }
 
-        public Ast_Struct(Token token): base(token)
+        public Ast_Struct(Token token) : base(token)
         {
             Type = AstType.Struct;
             StructType = token.Lexeme;
-            StructScope = new Ast_Scope(token.Lexeme, null);
+            StructScope = new Ast_Scope(token.Lexeme, null)
+            {
+                CanSearchUp = false
+            };
         }
 
         public Ast_Struct Clone(Token token)
         {
-            var str = new Ast_Struct(token) {
-                StructType = Name
-            };
-            foreach(var item in StructScope?.Variables)
+            var str = new Ast_Struct(token)
             {
-                if (item.Type == AstType.Function || item.Type == AstType.Procedure)
+                StructType = Name,
+                StructScope = StructScope.Parent.CreateChild(token.Lexeme)
+            };
+            StructScope.CanSearchUp = false;
+            foreach (var item in StructScope?.Variables)
+            {
+                if (item.Value.Type is ValueType.Function or ValueType.Procedure)
                 {
                     str.StructScope.Variables.Add(item);
                 }
                 else if (item.Type == AstType.Variable)
                 {
-                    var v = new Ast_Variable(item.Token);
-                    str.StructScope.Variables.Add(v);
+                    var value = new VT_Any();
+                    if (item.Value.Type == ValueType.Array)
+                    {
+                        value.Value = new List<VT_Any>();
+                        value.Type = ValueType.Array;
+                    }
+                    else if (item.Value.Type == ValueType.Record)
+                    {
+                        value.Value = new Dictionary<string, VT_Any>();
+                        value.Type = ValueType.Record;
+                    }
+                    else if (item.Value.Type == ValueType.Params)
+                    {
+                        value.Value = new List<VT_Any>();
+                        value.Type = ValueType.Params;
+                    }
+                    else
+                    {
+                        value.Value = item.Value.Value;
+                        value.Type = item.Value.Type;
+                    }
+                    _ = str.StructScope.Variables.Append(item.Name, value.Value);
                 }
                 else
                 {
                     throw new RuntimeError(token, $"Unknown type discovered: {item.Type}");
                 }
             }
-            return str;            
+            return str;
         }
 
         public override string ToString()
@@ -58,19 +80,20 @@ namespace Language
             structVar.DoSetValue(this);
             scope.Variables.Add(structVar);
             StructScope = scope.CreateChild(Name);
-            foreach(var ast in Block)
+            StructScope.CanSearchUp = false;
+            foreach (var ast in Block)
             {
                 if (ast.Type == AstType.Variable)
                 {
                     StructScope.Variables.Add(ast as Ast_Variable);
                 }
-                else if (ast.Type == AstType.Assign|| ast.Type == AstType.Function ||ast.Type == AstType.Procedure)
+                else if (ast.Type == AstType.Assign || ast.Type == AstType.Function || ast.Type == AstType.Procedure)
                 {
                     ast.Execute(StructScope);
                 }
                 else
                 {
-                    throw new RuntimeError(ast.Token,$"Invalid struct item type: {ast.Type}.");
+                    throw new RuntimeError(ast.Token, $"Invalid struct item type: {ast.Type}.");
                 }
             }
             return false;